requirements.txt
absl-py==1.2.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
astroid==2.5.6
asttokens==2.0.5
async-generator==1.10
attrs==21.4.0
Automat==20.2.0
autopep8==1.6.0
azure-cognitiveservices-speech==1.23.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
blurhash==1.1.4
boost==0.1
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
chromedriver-py==104.0.5112.29
cmake==3.22.5
colorama==0.4.4
comtypes==1.1.14
constantly==15.1.0
cryptography==37.0.2
cssselect==1.1.0
cycler==0.11.0
deap==1.3.1
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
Django==3.2.15
django-appconf==1.0.5
django-compressor==4.1
django-crispy-forms==1.14.0
django-debug-toolbar==3.7.0
django-environ==0.9.0
django-etc==1.3.1
django-extensions==3.2.1
django-hitcount==1.3.5
django-multi-captcha-admin==2.0.0
django-ranged-response==0.2.0
django-robots==5.0
django-simple-captcha==0.5.17
django-summernote==0.8.20.0
django-taggit==3.0.0
django-toast-messages==0.1
EasyTkinter==1.1.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fake-useragent==0.1.11
fastjsonschema==2.15.3
filelock==3.6.0
fonttools==4.34.4
graphviz==0.20.1
greenlet==1.1.2
h11==0.13.0
html5lib==1.1
hyperlink==21.0.0
idna==2.10
igraph==0.10.2
importlib-metadata==4.11.3
importlib-resources==5.7.1
imutils==0.5.4
incremental==21.3.0
instaloader==4.8.1
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
isort==5.8.0
itemadapter==0.5.0
itemloaders==1.0.4
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.0
joblib==1.1.0
jsonschema==4.5.1
jupyter==1.0.0
jupyter-client==7.3.1
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kiwisolver==1.4.4
lazy-object-proxy==1.6.0
lxml==4.8.0
MarkupSafe==2.1.1
Mastodon.py==1.5.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mccabe==0.6.1
media==0.1.6
mediapipe==0.8.10.1
mistune==0.8.4
mlxtend==0.21.0
mysql-connector-python==8.0.27
nbclient==0.6.2
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.3
notebook==6.4.11
numpy==1.22.3
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
openpyxl==3.0.9
outcome==1.1.0
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parse==1.19.0
parsel==1.6.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.1
pipenv==2022.5.2
platformdirs==2.5.2
pony==0.7.14
prometheus-client==0.14.1
prompt-toolkit==3.0.29
Protego==0.2.1
protobuf==3.20.1
psutil==5.9.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
PyDispatcher==2.0.5
pydotplus==2.0.2
pyee==8.2.2
pygame==2.1.2
Pygments==2.12.0
pylint==2.8.2
pyOpenSSL==22.0.0
pyparsing==3.0.8
pypiwin32==223
pyppeteer==1.0.2
PyQt5==5.12.2
PyQt5-sip==12.10.1
pyquery==1.4.3
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-magic==0.4.27
pyttsx3==2.90
pytz==2022.1
pywin32==304
pywinpty==2.0.5
pyzmq==22.3.0
qtconsole==5.3.0
QtPy==2.1.0
queuelib==1.6.2
qurl==0.0.1
rcssmin==1.1.0
requests==2.25.1
requests-file==1.5.1
requests-html==0.10.0
retry-requests==1.0.1
rjsmin==1.2.0
scikit-learn==1.1.2
scipy==1.9.0
Scrapy==2.6.1
seaborn==0.11.2
selenium==4.1.5
Send2Trash==1.8.0
service-identity==21.1.0
six==1.16.0
sklearn==0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.39
sqlparse==0.4.2
stack-data==0.2.0
terminado==0.13.3
texttable==1.6.4
threadpoolctl==3.1.0
tinycss2==1.1.1
tldextract==3.3.0
toml==0.10.2
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
trio==0.20.0
trio-websocket==0.9.2
Twisted==22.4.0
twisted-iocpsupport==1.0.2
typing_extensions==4.2.0
unicode==2.8
Unidecode==1.3.4
urllib3==1.26.4
virtualenv==20.14.1
virtualenv-clone==0.5.7
w3lib==1.22.0
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.3
widgetsnbextension==3.6.0
wrapt==1.12.1
wsproto==1.1.0
XlsxWriter==3.0.3
zipp==3.8.0
zope.interface==5.4.0
